+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `textColor` property on TextField ([#133](https://github.com/square/Blueprint/pull/133)).
+
 ### Removed
 
 ### Changed